requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.2.0
Django==5.0.7
django-adminlte2==0.4.3
django-ckeditor==6.7.0
django-js-asset==2.2.0
gunicorn==21.2.0
idna==3.10
Pillow==10.4.0
psycopg[binary,pool]==3.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
stripe==10.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
